requirements.txt
cryptography>=3.4.7
requests>=2.26.0
websocket-client>=1.2.1
python-dotenv>=0.19.0
asyncio>=3.4.3
typing-extensions>=4.0.1
openai>=0.27.2